# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "KParser API",
    "description": "Документация к АПИ парсер одного не безизвестного сайта",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/auth/token": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Токен",
        "description": "Получить новый токен",
        "parameters": [
          {
            "name": "partner",
            "description": "Токен партнера",
            "required": true,
            "in": "query",
            "type": "string"
          },
          {
            "name": "date_expire",
            "description": "Дата окончания работы токена, по умолчанию 1мес, максимум год",
            "required": false,
            "schema": {
              "type": "string",
              "example": "10.10.2010"
            },
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Новый токен пользователя. Если токен не авторизуют в течении 24 часов, то он становится недействительным:",
            "schema": {
              "type": "object",
              "properties": {
                "user_token": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/wa-dictionaries/big-cities": {
      "get": {
        "tags": [
          "Справочники (Необходима авторизация)"
        ],
        "summary": "Города",
        "description": "Получить список крупных городов",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список городов:",
            "schema": {
              "$ref": "#/definitions/BigCities"
            }
          }
        }
      }
    },
    "/wa-dictionaries/city-districts/{cityid}": {
      "get": {
        "tags": [
          "Справочники (Необходима авторизация)"
        ],
        "summary": "Районы",
        "description": "Получить список районов города",
        "parameters": [
          {
            "name": "cityid",
            "in": "path",
            "required": true,
            "description": "ID Города"
          }
        ],
        "responses": {
          "200": {
            "description": "Список районов",
            "schema": {
              "$ref": "#/definitions/AllCityDistricts"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "BigCity": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Алматы"
        },
        "alias": {
          "type": "string",
          "example": "Almaty"
        },
        "id": {
          "type": "number",
          "example": 2
        }
      }
    },
    "BigCities": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/BigCity"
      }
    },
    "CityDistrict": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 3
        },
        "alias": {
          "type": "string",
          "example": "almaty-alatauskij"
        },
        "name": {
          "type": "string",
          "example": "Алатауский р-н"
        },
        "lat": {
          "type": "number",
          "example": 43.2813
        },
        "lon": {
          "type": "number",
          "example": 76.852
        },
        "zoom": {
          "type": "number",
          "example": 13
        },
        "hasChildren": {
          "type": "boolean",
          "example": true
        },
        "hasComplexes": {
          "type": "boolean",
          "example": false
        },
        "isBigCity": {
          "type": "boolean",
          "example": false
        },
        "type": {
          "type": "string",
          "example": "district"
        },
        "parentId": {
          "type": "number",
          "example": 2
        },
        "level": {
          "type": "number",
          "example": 3
        }
      }
    },
    "AllCityDistricts": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CityDistrict"
          }
        },
        "regionCount": {
          "type": "number",
          "example": 9
        },
        "parent": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 2
            },
            "lon": {
              "type": "number",
              "example": 76.9129
            },
            "zoom": {
              "type": "number",
              "example": 11
            },
            "type": {
              "type": "string",
              "example": "city"
            },
            "hasComplexes": {
              "type": "boolean",
              "example": true
            },
            "isSelectable": {
              "type": "boolean",
              "example": true
            }
          }
        }
      }
    },
    "Text": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string",
          "example": "test"
        }
      }
    },
    "Changes": {
      "type": "object",
      "properties": {
        "changes": {
          "type": "object",
          "properties": {
            "text": {
              "type": "string",
              "example": "test"
            },
            "done": {
              "type": "boolean",
              "example": true
            }
          }
        }
      }
    }
  }
}